refactor: reorg schema and loader related code
