+++ src/routes/loader.ts
@@ -1,6 +1,26 @@
 import { type QueryClient } from '@tanstack/react-query';
 import { type LoaderFunctionArgs } from 'react-router';
-import { allProgramsKeys } from '~/routes/use-filtered-programs';
+import { fetchApi } from '~/api/fetch';
+import { type Program } from './schema';
+
+export const fetchAllPrograms = (): Promise<Program[]> => {
+  return fetchApi('/wp-json/nutrition-navigator/v1/programs').then((res) =>
+    res.json()
+  );
+};
+
+export const allProgramsKeys = {
+  all: ['allPrograms'] as const,
+  filtered: (filters: {
+    address: string;
+    regions: Array<string>;
+    programTypes: Array<string>;
+    languages: Array<string>;
+    venues: Array<string>;
+    audiences: Array<string>;
+    organizationName: string;
+  }) => [...allProgramsKeys.all, { filters }] as const
+};
 
 export const loader =
   (queryClient: QueryClient) =>
